Update is_unique.py
With the brute force method, the first character in the string will be compared with the next character until the last character. Then the second character will be compared but started from the next character, not from the first character while skipping itself.  This iteration continues until the second last character which will be compared with the last one.